pyproject.toml
[project]
name = "seo-crawler"
version = "0.1.0"
description = "Local SEO crawler backend - ETAP 2"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.13.2",
  "fastapi>=0.115.0",
  "pydantic-settings>=2.4.0",
  "psycopg[binary]>=3.2.1",
  "scrapy>=2.11.2",
  "sqlalchemy>=2.0.32",
  "tldextract>=5.1.2",
  "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0",
  "pytest>=8.3.2",
]

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
markers = [
  "postgres_smoke: smoke tests that require running PostgreSQL instance",
]
